+++ src/core/gps/tok.h
@@ -12,6 +12,7 @@
 #define __NMEA_TOK_H__
 
 #include "config.h"
+#include "stdlib.h"
 
 #ifdef  __cplusplus
 extern "C"
@@ -19,7 +20,7 @@ extern "C"
 #endif
 
   int     nmea_calc_crc( const char *buff, int buff_sz );
-  int     nmea_atoi( const char *str, int str_sz, int radix );
+  int     nmea_atoi( const char *str, size_t str_sz, int radix );
   double  nmea_atof( const char *str, int str_sz );
   int     nmea_printf( char *buff, int buff_sz, const char *format, ... );
   int     nmea_scanf( const char *buff, int buff_sz, const char *format, ... );
